removing the warning and forcing TCP port to default when changing of protocol
not all users may know what is the default port for FTP or SFTP/SSH
and so forcing it to default upon change of protocol seems a better
solution as it does not make sense that upon application protocol change the
TCP port stays the same. This will still be possible to set it back to the
same value after changing from FTP to SFTP or vice versa.