avoiding updating an html_form_input_file with a new repo info upon bibliotheque change when this component is not visible
this update through the actor::on_event() method and was driving html_web_user_interaction
subcomponent to get visible to manage a thread spawns to create the new repo. Though if the
component that had reacted to through its on_event() is not visibile the html_web_user_interaction
cannot work as expected. The solution is to flag the need of repo update from the on_event()
method and do the repo update and html_web_user_interaction + thread execution from the
inherited_get_body_part().