fixed html_options_read delayed changed event
when calling load_json, the ref_entrepot parameters are set but the
entrepot is just flagged as to be re-generated. This occurs outside load_json
when the html_options_read is asked to provide a body part (inherited_get_body_part)
and at that time only this generates an change event, so leads the html_bibliotheque
to warn about the just loaded html_options_read to have unsaved changes.
To fix that, a bool value is stored in html_options_read and set to true when load_json()
is called it is cleared after the ref_entrepot is re-generated and prevent in that
context the change event to be generated