enforcing control between libdar thread and webdar through the web_user_interaction class
the signal() could awake the libdar thread pending on wait() of the same libthreadar::condition
so long after (several seconds) [only observered under voidLinux+musl] that this triggered
self reported bug condition. It could also be the point that the signal() missed awaking
the libdar thread suspending on wait()... ? Anyway it was necessary to improve thread
synchronization.